Crate stm32f4

Peripheral access API for STM32F4 microcontrollers (generated using svd2rust 0.17.0)

You can find an overview of the API here: svd2rust/#peripheral-api

For more details see the README here: stm32-rs

This crate supports all STM32F4 devices; for the complete list please see: stm32f4

Due to doc build limitations, not all devices may be shown on docs.rs; a representative few have been selected instead. For a complete list of available registers and fields see: stm32-rs Device Coverage

Modules

stm32f405

Structs

R
Register/field reader

Reg
This structure provides volatile access to register

W
Register writer

Enums

Variant
Used if enumerated values cover not the whole range

Traits

Readable
This trait shows that register has read method

ResetValue
Reset value of the register

Writable
This trait shows that register has write, write_with_zero and reset method