Type Definition stm32f4::stm32f405::adc1::cr2::EXTSEL_R

type EXTSEL_R = R<u8, EXTSEL_A>;

Reader of field EXTSEL

Implementations

impl EXTSEL_R

pub fn variant(&self) -> Variant<u8, EXTSEL_A>

Get enumerated values variant

pub fn is_tim1cc1(&self) -> bool

Checks if the value of the field is TIM1CC1

pub fn is_tim1cc2(&self) -> bool

Checks if the value of the field is TIM1CC2

pub fn is_tim1cc3(&self) -> bool

Checks if the value of the field is TIM1CC3

pub fn is_tim2cc2(&self) -> bool

Checks if the value of the field is TIM2CC2

pub fn is_tim2cc3(&self) -> bool

Checks if the value of the field is TIM2CC3

pub fn is_tim2cc4(&self) -> bool

Checks if the value of the field is TIM2CC4

pub fn is_tim2trgo(&self) -> bool

Checks if the value of the field is TIM2TRGO