Type Definition stm32f4::stm32f405::adc1::cr2::JEXTEN_R

type JEXTEN_R = R<u8, JEXTEN_A>;

Reader of field JEXTEN

Implementations

impl JEXTEN_R

pub fn variant(&self) -> JEXTEN_A

Get enumerated values variant

pub fn is_disabled(&self) -> bool

Checks if the value of the field is DISABLED

pub fn is_rising_edge(&self) -> bool

Checks if the value of the field is RISINGEDGE

pub fn is_falling_edge(&self) -> bool

Checks if the value of the field is FALLINGEDGE

pub fn is_both_edges(&self) -> bool

Checks if the value of the field is BOTHEDGES