Type Definition stm32f4::stm32f405::adc1::smpr1::SMP18_R

type SMP18_R = R<u8, SMP18_A>;

Reader of field SMP18

Implementations

impl SMP18_R

pub fn variant(&self) -> SMP18_A

Get enumerated values variant

pub fn is_cycles3(&self) -> bool

Checks if the value of the field is CYCLES3

pub fn is_cycles15(&self) -> bool

Checks if the value of the field is CYCLES15

pub fn is_cycles28(&self) -> bool

Checks if the value of the field is CYCLES28

pub fn is_cycles56(&self) -> bool

Checks if the value of the field is CYCLES56

pub fn is_cycles84(&self) -> bool

Checks if the value of the field is CYCLES84

pub fn is_cycles112(&self) -> bool

Checks if the value of the field is CYCLES112

pub fn is_cycles144(&self) -> bool

Checks if the value of the field is CYCLES144

pub fn is_cycles480(&self) -> bool

Checks if the value of the field is CYCLES480