Module stm32f4::stm32f405::adc_common

Common ADC registers

Modules

ccr
ADC common control register

cdr
ADC common regular data register for dual and triple modes

csr
ADC Common status register

Structs

RegisterBlock
Register block

Type Definitions

CCR
ADC common control register

CDR
ADC common regular data register for dual and triple modes

CSR
ADC Common status register