Type Definition stm32f4::stm32f405::adc_common::ccr::MULTI_R

type MULTI_R = R<u8, MULTI_A>;

Reader of field MULTI

Implementations

impl MULTI_R

pub fn variant(&self) -> Variant<u8, MULTI_A>

Get enumerated values variant

pub fn is_independent(&self) -> bool

Checks if the value of the field is INDEPENDENT

pub fn is_dual_rj(&self) -> bool

Checks if the value of the field is DUALRJ

pub fn is_dual_ra(&self) -> bool

Checks if the value of the field is DUALRA

pub fn is_dual_j(&self) -> bool

Checks if the value of the field is DUALJ

pub fn is_dual_r(&self) -> bool

Checks if the value of the field is DUALR

pub fn is_dual_i(&self) -> bool

Checks if the value of the field is DUALI

pub fn is_dual_a(&self) -> bool

Checks if the value of the field is DUALA

pub fn is_triple_rj(&self) -> bool

Checks if the value of the field is TRIPLERJ

pub fn is_triple_ra(&self) -> bool

Checks if the value of the field is TRIPLERA

pub fn is_triple_j(&self) -> bool

Checks if the value of the field is TRIPLEJ

pub fn is_triple_r(&self) -> bool

Checks if the value of the field is TRIPLER

pub fn is_triple_i(&self) -> bool

Checks if the value of the field is TRIPLEI

pub fn is_triple_a(&self) -> bool

Checks if the value of the field is TRIPLEA