Type Definition stm32f4::stm32f405::can1::esr::LEC_R

type LEC_R = R<u8, LEC_A>;

Reader of field LEC

Implementations

impl LEC_R

pub fn variant(&self) -> LEC_A

Get enumerated values variant

pub fn is_no_error(&self) -> bool

Checks if the value of the field is NOERROR

pub fn is_stuff(&self) -> bool

Checks if the value of the field is STUFF

pub fn is_form(&self) -> bool

Checks if the value of the field is FORM

pub fn is_ack(&self) -> bool

Checks if the value of the field is ACK

pub fn is_bit_recessive(&self) -> bool

Checks if the value of the field is BITRECESSIVE

pub fn is_bit_dominant(&self) -> bool

Checks if the value of the field is BITDOMINANT

pub fn is_crc(&self) -> bool

Checks if the value of the field is CRC

pub fn is_custom(&self) -> bool

Checks if the value of the field is CUSTOM