Type Definition stm32f4::stm32f405::dma2::lifcr::W

type W = W<u32, LIFCR>;

Writer for register LIFCR

Implementations

impl W

pub fn ctcif3(&mut self) -> CTCIF3_W<'_>

Bit 27 - Stream x clear transfer complete interrupt flag (x = 3..0)

pub fn chtif3(&mut self) -> CHTIF3_W<'_>

Bit 26 - Stream x clear half transfer interrupt flag (x = 3..0)

pub fn cteif3(&mut self) -> CTEIF3_W<'_>

Bit 25 - Stream x clear transfer error interrupt flag (x = 3..0)

pub fn cdmeif3(&mut self) -> CDMEIF3_W<'_>

Bit 24 - Stream x clear direct mode error interrupt flag (x = 3..0)

pub fn cfeif3(&mut self) -> CFEIF3_W<'_>

Bit 22 - Stream x clear FIFO error interrupt flag (x = 3..0)

pub fn ctcif2(&mut self) -> CTCIF2_W<'_>

Bit 21 - Stream x clear transfer complete interrupt flag (x = 3..0)

pub fn chtif2(&mut self) -> CHTIF2_W<'_>

Bit 20 - Stream x clear half transfer interrupt flag (x = 3..0)

pub fn cteif2(&mut self) -> CTEIF2_W<'_>

Bit 19 - Stream x clear transfer error interrupt flag (x = 3..0)

pub fn cdmeif2(&mut self) -> CDMEIF2_W<'_>

Bit 18 - Stream x clear direct mode error interrupt flag (x = 3..0)

pub fn cfeif2(&mut self) -> CFEIF2_W<'_>

Bit 16 - Stream x clear FIFO error interrupt flag (x = 3..0)

pub fn ctcif1(&mut self) -> CTCIF1_W<'_>

Bit 11 - Stream x clear transfer complete interrupt flag (x = 3..0)

pub fn chtif1(&mut self) -> CHTIF1_W<'_>

Bit 10 - Stream x clear half transfer interrupt flag (x = 3..0)

pub fn cteif1(&mut self) -> CTEIF1_W<'_>

Bit 9 - Stream x clear transfer error interrupt flag (x = 3..0)

pub fn cdmeif1(&mut self) -> CDMEIF1_W<'_>

Bit 8 - Stream x clear direct mode error interrupt flag (x = 3..0)

pub fn cfeif1(&mut self) -> CFEIF1_W<'_>

Bit 6 - Stream x clear FIFO error interrupt flag (x = 3..0)

pub fn ctcif0(&mut self) -> CTCIF0_W<'_>

Bit 5 - Stream x clear transfer complete interrupt flag (x = 3..0)

pub fn chtif0(&mut self) -> CHTIF0_W<'_>

Bit 4 - Stream x clear half transfer interrupt flag (x = 3..0)

pub fn cteif0(&mut self) -> CTEIF0_W<'_>

Bit 3 - Stream x clear transfer error interrupt flag (x = 3..0)

pub fn cdmeif0(&mut self) -> CDMEIF0_W<'_>

Bit 2 - Stream x clear direct mode error interrupt flag (x = 3..0)

pub fn cfeif0(&mut self) -> CFEIF0_W<'_>

Bit 0 - Stream x clear FIFO error interrupt flag (x = 3..0)