Type Definition stm32f4::stm32f405::gpioa::afrh::W

type W = W<u32, AFRH>;

Writer for register AFRH

Implementations

impl W

pub fn afrh15(&mut self) -> AFRH15_W<'_>

Bits 28:31 - Alternate function selection for port x bit y (y = 8..15)

pub fn afrh14(&mut self) -> AFRH14_W<'_>

Bits 24:27 - Alternate function selection for port x bit y (y = 8..15)

pub fn afrh13(&mut self) -> AFRH13_W<'_>

Bits 20:23 - Alternate function selection for port x bit y (y = 8..15)

pub fn afrh12(&mut self) -> AFRH12_W<'_>

Bits 16:19 - Alternate function selection for port x bit y (y = 8..15)

pub fn afrh11(&mut self) -> AFRH11_W<'_>

Bits 12:15 - Alternate function selection for port x bit y (y = 8..15)

pub fn afrh10(&mut self) -> AFRH10_W<'_>

Bits 8:11 - Alternate function selection for port x bit y (y = 8..15)

pub fn afrh9(&mut self) -> AFRH9_W<'_>

Bits 4:7 - Alternate function selection for port x bit y (y = 8..15)

pub fn afrh8(&mut self) -> AFRH8_W<'_>

Bits 0:3 - Alternate function selection for port x bit y (y = 8..15)