Type Definition stm32f4::stm32f405::gpiob::afrl::AFRL7_R

type AFRL7_R = R<u8, AFRL7_A>;

Reader of field AFRL7

Implementations

impl AFRL7_R

pub fn variant(&self) -> AFRL7_A

Get enumerated values variant

pub fn is_af0(&self) -> bool

Checks if the value of the field is AF0

pub fn is_af1(&self) -> bool

Checks if the value of the field is AF1

pub fn is_af2(&self) -> bool

Checks if the value of the field is AF2

pub fn is_af3(&self) -> bool

Checks if the value of the field is AF3

pub fn is_af4(&self) -> bool

Checks if the value of the field is AF4

pub fn is_af5(&self) -> bool

Checks if the value of the field is AF5

pub fn is_af6(&self) -> bool

Checks if the value of the field is AF6

pub fn is_af7(&self) -> bool

Checks if the value of the field is AF7

pub fn is_af8(&self) -> bool

Checks if the value of the field is AF8

pub fn is_af9(&self) -> bool

Checks if the value of the field is AF9

pub fn is_af10(&self) -> bool

Checks if the value of the field is AF10

pub fn is_af11(&self) -> bool

Checks if the value of the field is AF11

pub fn is_af12(&self) -> bool

Checks if the value of the field is AF12

pub fn is_af13(&self) -> bool

Checks if the value of the field is AF13

pub fn is_af14(&self) -> bool

Checks if the value of the field is AF14

pub fn is_af15(&self) -> bool

Checks if the value of the field is AF15