Type Definition stm32f4::stm32f405::gpiob::idr::IDR15_R

type IDR15_R = R<bool, IDR15_A>;

Reader of field IDR15

Implementations

impl IDR15_R

pub fn variant(&self) -> IDR15_A

Get enumerated values variant

pub fn is_high(&self) -> bool

Checks if the value of the field is HIGH

pub fn is_low(&self) -> bool

Checks if the value of the field is LOW