Type Definition stm32f4::stm32f405::gpioi::afrh::R

type R = R<u32, AFRH>;

Reader of register AFRH

Implementations

impl R

pub fn afrh15(&self) -> AFRH15_R

Bits 28:31 - Alternate function selection for port x bit y (y = 8..15)

pub fn afrh14(&self) -> AFRH14_R

Bits 24:27 - Alternate function selection for port x bit y (y = 8..15)

pub fn afrh13(&self) -> AFRH13_R

Bits 20:23 - Alternate function selection for port x bit y (y = 8..15)

pub fn afrh12(&self) -> AFRH12_R

Bits 16:19 - Alternate function selection for port x bit y (y = 8..15)

pub fn afrh11(&self) -> AFRH11_R

Bits 12:15 - Alternate function selection for port x bit y (y = 8..15)

pub fn afrh10(&self) -> AFRH10_R

Bits 8:11 - Alternate function selection for port x bit y (y = 8..15)

pub fn afrh9(&self) -> AFRH9_R

Bits 4:7 - Alternate function selection for port x bit y (y = 8..15)

pub fn afrh8(&self) -> AFRH8_R

Bits 0:3 - Alternate function selection for port x bit y (y = 8..15)