Type Definition stm32f4::stm32f405::iwdg::pr::PR_R

type PR_R = R<u8, PR_A>;

Reader of field PR

Implementations

impl PR_R

pub fn variant(&self) -> PR_A

Get enumerated values variant

pub fn is_divide_by4(&self) -> bool

Checks if the value of the field is DIVIDEBY4

pub fn is_divide_by8(&self) -> bool

Checks if the value of the field is DIVIDEBY8

pub fn is_divide_by16(&self) -> bool

Checks if the value of the field is DIVIDEBY16

pub fn is_divide_by32(&self) -> bool

Checks if the value of the field is DIVIDEBY32

pub fn is_divide_by64(&self) -> bool

Checks if the value of the field is DIVIDEBY64

pub fn is_divide_by128(&self) -> bool

Checks if the value of the field is DIVIDEBY128

pub fn is_divide_by256(&self) -> bool

Checks if the value of the field is DIVIDEBY256

pub fn is_divide_by256bis(&self) -> bool

Checks if the value of the field is DIVIDEBY256BIS