Type Definition stm32f4::stm32f405::rcc::cfgr::HPRE_R

type HPRE_R = R<u8, HPRE_A>;

Reader of field HPRE

Implementations

impl HPRE_R

pub fn variant(&self) -> Variant<u8, HPRE_A>

Get enumerated values variant

pub fn is_div1(&self) -> bool

Checks if the value of the field is DIV1

pub fn is_div2(&self) -> bool

Checks if the value of the field is DIV2

pub fn is_div4(&self) -> bool

Checks if the value of the field is DIV4

pub fn is_div8(&self) -> bool

Checks if the value of the field is DIV8

pub fn is_div16(&self) -> bool

Checks if the value of the field is DIV16

pub fn is_div64(&self) -> bool

Checks if the value of the field is DIV64

pub fn is_div128(&self) -> bool

Checks if the value of the field is DIV128

pub fn is_div256(&self) -> bool

Checks if the value of the field is DIV256

pub fn is_div512(&self) -> bool

Checks if the value of the field is DIV512