Type Definition stm32f4::stm32f405::rcc::cfgr::MCO2PRE_R

type MCO2PRE_R = R<u8, MCO2PRE_A>;

Reader of field MCO2PRE

Implementations

impl MCO2PRE_R

pub fn variant(&self) -> Variant<u8, MCO2PRE_A>

Get enumerated values variant

pub fn is_div1(&self) -> bool

Checks if the value of the field is DIV1

pub fn is_div2(&self) -> bool

Checks if the value of the field is DIV2

pub fn is_div3(&self) -> bool

Checks if the value of the field is DIV3

pub fn is_div4(&self) -> bool

Checks if the value of the field is DIV4

pub fn is_div5(&self) -> bool

Checks if the value of the field is DIV5