Type Definition stm32f4::stm32f405::rcc::cfgr::PPRE2_R

type PPRE2_R = R<u8, PPRE2_A>;

Reader of field PPRE2

Implementations

impl PPRE2_R

pub fn variant(&self) -> Variant<u8, PPRE2_A>

Get enumerated values variant

pub fn is_div1(&self) -> bool

Checks if the value of the field is DIV1

pub fn is_div2(&self) -> bool

Checks if the value of the field is DIV2

pub fn is_div4(&self) -> bool

Checks if the value of the field is DIV4

pub fn is_div8(&self) -> bool

Checks if the value of the field is DIV8

pub fn is_div16(&self) -> bool

Checks if the value of the field is DIV16