Type Definition stm32f4::stm32f405::rcc::cfgr::SW_R

type SW_R = R<u8, SW_A>;

Reader of field SW

Implementations

impl SW_R

pub fn variant(&self) -> Variant<u8, SW_A>

Get enumerated values variant

pub fn is_hsi(&self) -> bool

Checks if the value of the field is HSI

pub fn is_hse(&self) -> bool

Checks if the value of the field is HSE

pub fn is_pll(&self) -> bool

Checks if the value of the field is PLL