Type Definition stm32f4::stm32f405::sai1::ch::cr1::DS_R

type DS_R = R<u8, DS_A>;

Reader of field DS

Implementations

impl DS_R

pub fn variant(&self) -> Variant<u8, DS_A>

Get enumerated values variant

pub fn is_bit8(&self) -> bool

Checks if the value of the field is BIT8

pub fn is_bit10(&self) -> bool

Checks if the value of the field is BIT10

pub fn is_bit16(&self) -> bool

Checks if the value of the field is BIT16

pub fn is_bit20(&self) -> bool

Checks if the value of the field is BIT20

pub fn is_bit24(&self) -> bool

Checks if the value of the field is BIT24

pub fn is_bit32(&self) -> bool

Checks if the value of the field is BIT32