Type Definition stm32f4::stm32f405::sai1::ch::frcr::FSPOL_R

type FSPOL_R = R<bool, FSPOL_A>;

Reader of field FSPOL

Implementations

impl FSPOL_R

pub fn variant(&self) -> FSPOL_A

Get enumerated values variant

pub fn is_falling_edge(&self) -> bool

Checks if the value of the field is FALLINGEDGE

pub fn is_rising_edge(&self) -> bool

Checks if the value of the field is RISINGEDGE