Type Definition stm32f4::stm32f405::spi1::cr1::BR_R

type BR_R = R<u8, BR_A>;

Reader of field BR

Implementations

impl BR_R

pub fn variant(&self) -> BR_A

Get enumerated values variant

pub fn is_div2(&self) -> bool

Checks if the value of the field is DIV2

pub fn is_div4(&self) -> bool

Checks if the value of the field is DIV4

pub fn is_div8(&self) -> bool

Checks if the value of the field is DIV8

pub fn is_div16(&self) -> bool

Checks if the value of the field is DIV16

pub fn is_div32(&self) -> bool

Checks if the value of the field is DIV32

pub fn is_div64(&self) -> bool

Checks if the value of the field is DIV64

pub fn is_div128(&self) -> bool

Checks if the value of the field is DIV128

pub fn is_div256(&self) -> bool

Checks if the value of the field is DIV256