Type Definition stm32f4::stm32f405::spi1::cr2::ERRIE_R

type ERRIE_R = R<bool, ERRIE_A>;

Reader of field ERRIE

Implementations

impl ERRIE_R

pub fn variant(&self) -> ERRIE_A

Get enumerated values variant

pub fn is_masked(&self) -> bool

Checks if the value of the field is MASKED

pub fn is_not_masked(&self) -> bool

Checks if the value of the field is NOTMASKED