Type Definition stm32f4::stm32f405::tim3::ccmr1_input::IC1F_R

type IC1F_R = R<u8, IC1F_A>;

Reader of field IC1F

Implementations

impl IC1F_R

pub fn variant(&self) -> IC1F_A

Get enumerated values variant

pub fn is_no_filter(&self) -> bool

Checks if the value of the field is NOFILTER

pub fn is_fck_int_n2(&self) -> bool

Checks if the value of the field is FCK_INT_N2

pub fn is_fck_int_n4(&self) -> bool

Checks if the value of the field is FCK_INT_N4

pub fn is_fck_int_n8(&self) -> bool

Checks if the value of the field is FCK_INT_N8

pub fn is_fdts_div2_n6(&self) -> bool

Checks if the value of the field is FDTS_DIV2_N6

pub fn is_fdts_div2_n8(&self) -> bool

Checks if the value of the field is FDTS_DIV2_N8

pub fn is_fdts_div4_n6(&self) -> bool

Checks if the value of the field is FDTS_DIV4_N6

pub fn is_fdts_div4_n8(&self) -> bool

Checks if the value of the field is FDTS_DIV4_N8

pub fn is_fdts_div8_n6(&self) -> bool

Checks if the value of the field is FDTS_DIV8_N6

pub fn is_fdts_div8_n8(&self) -> bool

Checks if the value of the field is FDTS_DIV8_N8

pub fn is_fdts_div16_n5(&self) -> bool

Checks if the value of the field is FDTS_DIV16_N5

pub fn is_fdts_div16_n6(&self) -> bool

Checks if the value of the field is FDTS_DIV16_N6

pub fn is_fdts_div16_n8(&self) -> bool

Checks if the value of the field is FDTS_DIV16_N8

pub fn is_fdts_div32_n5(&self) -> bool

Checks if the value of the field is FDTS_DIV32_N5

pub fn is_fdts_div32_n6(&self) -> bool

Checks if the value of the field is FDTS_DIV32_N6

pub fn is_fdts_div32_n8(&self) -> bool

Checks if the value of the field is FDTS_DIV32_N8