Type Definition stm32f4::stm32f405::tim3::smcr::SMS_R

type SMS_R = R<u8, SMS_A>;

Reader of field SMS

Implementations

impl SMS_R

pub fn variant(&self) -> SMS_A

Get enumerated values variant

pub fn is_disabled(&self) -> bool

Checks if the value of the field is DISABLED

pub fn is_encoder_mode_1(&self) -> bool

Checks if the value of the field is ENCODER_MODE_1

pub fn is_encoder_mode_2(&self) -> bool

Checks if the value of the field is ENCODER_MODE_2

pub fn is_encoder_mode_3(&self) -> bool

Checks if the value of the field is ENCODER_MODE_3

pub fn is_reset_mode(&self) -> bool

Checks if the value of the field is RESET_MODE

pub fn is_gated_mode(&self) -> bool

Checks if the value of the field is GATED_MODE

pub fn is_trigger_mode(&self) -> bool

Checks if the value of the field is TRIGGER_MODE

pub fn is_ext_clock_mode(&self) -> bool

Checks if the value of the field is EXT_CLOCK_MODE