Type Definition stm32f4::stm32f405::tim5::ccmr2_output::OC4M_R

type OC4M_R = R<u8, OC4M_A>;

Reader of field OC4M

Implementations

impl OC4M_R

pub fn variant(&self) -> OC4M_A

Get enumerated values variant

pub fn is_frozen(&self) -> bool

Checks if the value of the field is FROZEN

pub fn is_active_on_match(&self) -> bool

Checks if the value of the field is ACTIVEONMATCH

pub fn is_inactive_on_match(&self) -> bool

Checks if the value of the field is INACTIVEONMATCH

pub fn is_toggle(&self) -> bool

Checks if the value of the field is TOGGLE

pub fn is_force_inactive(&self) -> bool

Checks if the value of the field is FORCEINACTIVE

pub fn is_force_active(&self) -> bool

Checks if the value of the field is FORCEACTIVE

pub fn is_pwm_mode1(&self) -> bool

Checks if the value of the field is PWMMODE1

pub fn is_pwm_mode2(&self) -> bool

Checks if the value of the field is PWMMODE2