Type Definition stm32f4::stm32f405::tim5::cr2::MMS_R

type MMS_R = R<u8, MMS_A>;

Reader of field MMS

Implementations

impl MMS_R

pub fn variant(&self) -> MMS_A

Get enumerated values variant

pub fn is_reset(&self) -> bool

Checks if the value of the field is RESET

pub fn is_enable(&self) -> bool

Checks if the value of the field is ENABLE

pub fn is_update(&self) -> bool

Checks if the value of the field is UPDATE

pub fn is_compare_pulse(&self) -> bool

Checks if the value of the field is COMPAREPULSE

pub fn is_compare_oc1(&self) -> bool

Checks if the value of the field is COMPAREOC1

pub fn is_compare_oc2(&self) -> bool

Checks if the value of the field is COMPAREOC2

pub fn is_compare_oc3(&self) -> bool

Checks if the value of the field is COMPAREOC3

pub fn is_compare_oc4(&self) -> bool

Checks if the value of the field is COMPAREOC4