Type Definition stm32f4::stm32f405::usart1::cr2::STOP_R

type STOP_R = R<u8, STOP_A>;

Reader of field STOP

Implementations

impl STOP_R

pub fn variant(&self) -> STOP_A

Get enumerated values variant

pub fn is_stop1(&self) -> bool

Checks if the value of the field is STOP1

pub fn is_stop0p5(&self) -> bool

Checks if the value of the field is STOP0P5

pub fn is_stop2(&self) -> bool

Checks if the value of the field is STOP2

pub fn is_stop1p5(&self) -> bool

Checks if the value of the field is STOP1P5