Module stm32f4::stm32f411

Peripheral access API for STM32F411 microcontrollers (generated using svd2rust v0.19.0 ( ))

You can find an overview of the generated API here.

API features to be included in the next svd2rust release can be generated by cloning the svd2rust repository, checking out the above commit, and running cargo doc --open.

Re-exports

pub use self::Interrupt as interrupt;

Modules

adc1

Analog-to-digital converter

adc_common

ADC common registers

crc

Cryptographic processor

dbgmcu

Debug support

dma1

DMA controller

exti

External interrupt/event controller

flash

FLASH

fpu

Floting point unit

fpu_cpacr

Floating point unit CPACR

gpioa

General-purpose I/Os

gpiob

General-purpose I/Os

gpioh

General-purpose I/Os

i2c1

Inter-integrated circuit

iwdg

Independent watchdog

nvic_stir

Nested vectored interrupt controller

otg_fs_device

USB on the go full speed

otg_fs_global

USB on the go full speed

otg_fs_host

USB on the go full speed

otg_fs_pwrclk

USB on the go full speed

pwr

Power control

rcc

Reset and clock control

rtc

Real-time clock

scb_actrl

System control block ACTLR

sdio

Secure digital input/output interface

spi1

Serial peripheral interface

stk

SysTick timer

syscfg

System configuration controller

tim1

Advanced-timers

tim2

General purpose timers

tim3

General purpose timers

tim5

General-purpose-timers

tim9

General purpose timers

tim10

General-purpose-timers

tim11

General-purpose-timers

usart1

Universal synchronous asynchronous receiver transmitter

wwdg

Window watchdog

Structs

ADC1

Analog-to-digital converter

ADC_COMMON

ADC common registers

CBP

Cache and branch predictor maintenance operations

CPUID

CPUID

CRC

Cryptographic processor

CorePeripherals

Core peripherals

DBGMCU

Debug support

DCB

Debug Control Block

DMA1

DMA controller

DMA2

DMA controller

DWT

Data Watchpoint and Trace unit

EXTI

External interrupt/event controller

FLASH

FLASH

FPB

Flash Patch and Breakpoint unit

FPU

Floting point unit

FPU_CPACR

Floating point unit CPACR

GPIOA

General-purpose I/Os

GPIOB

General-purpose I/Os

GPIOC

General-purpose I/Os

GPIOD

General-purpose I/Os

GPIOE

General-purpose I/Os

GPIOH

General-purpose I/Os

I2C1

Inter-integrated circuit

I2C2

Inter-integrated circuit

I2C3

Inter-integrated circuit

I2S2EXT

Serial peripheral interface

I2S3EXT

Serial peripheral interface

ITM

Instrumentation Trace Macrocell

IWDG

Independent watchdog

MPU

Memory Protection Unit

NVIC

Nested Vector Interrupt Controller

NVIC_STIR

Nested vectored interrupt controller

OTG_FS_DEVICE

USB on the go full speed

OTG_FS_GLOBAL

USB on the go full speed

OTG_FS_HOST

USB on the go full speed

OTG_FS_PWRCLK

USB on the go full speed

PWR

Power control

Peripherals

All the peripherals

RCC

Reset and clock control

RTC

Real-time clock

SCB

System Control Block

SCB_ACTRL

System control block ACTLR

SDIO

Secure digital input/output interface

SPI1

Serial peripheral interface

SPI2

Serial peripheral interface

SPI3

Serial peripheral interface

SPI4

Serial peripheral interface

SPI5

Serial peripheral interface

STK

SysTick timer

SYSCFG

System configuration controller

SYST

SysTick: System Timer

TIM1

Advanced-timers

TIM2

General purpose timers

TIM3

General purpose timers

TIM4

General purpose timers

TIM5

General-purpose-timers

TIM8

Advanced-timers

TIM9

General purpose timers

TIM10

General-purpose-timers

TIM11

General-purpose-timers

TPIU

Trace Port Interface Unit

USART1

Universal synchronous asynchronous receiver transmitter

USART2

Universal synchronous asynchronous receiver transmitter

USART6

Universal synchronous asynchronous receiver transmitter

WWDG

Window watchdog

Enums

Interrupt

Enumeration of all the interrupts.

Constants

NVIC_PRIO_BITS

Number available in the NVIC for configuring priority

Attribute Macros

interrupt