Type Definition stm32f4::stm32f411::adc1::smpr1::SMPX_X_R

type SMPX_X_R = R<u32, SMPX_X_A>;

Reader of field SMPx_x

Implementations

impl SMPX_X_R

pub fn variant(&self) -> Variant<u32, SMPX_X_A>

Get enumerated values variant

pub fn is_cycles3(&self) -> bool

Checks if the value of the field is CYCLES3

pub fn is_cycles15(&self) -> bool

Checks if the value of the field is CYCLES15

pub fn is_cycles28(&self) -> bool

Checks if the value of the field is CYCLES28

pub fn is_cycles56(&self) -> bool

Checks if the value of the field is CYCLES56

pub fn is_cycles84(&self) -> bool

Checks if the value of the field is CYCLES84

pub fn is_cycles112(&self) -> bool

Checks if the value of the field is CYCLES112

pub fn is_cycles144(&self) -> bool

Checks if the value of the field is CYCLES144

pub fn is_cycles480(&self) -> bool

Checks if the value of the field is CYCLES480