Struct stm32f4::stm32f411::i2c1::fltr::DNF_W

pub struct DNF_W<'a> { /* fields omitted */ }

Field DNF writer - Digital noise filter

Implementations

impl<'a> DNF_W<'a>

pub fn variant(self, variant: DNF_A) -> &'a mut W

Writes variant to the field

pub fn no_filter(self) -> &'a mut W

Digital filter disabled

pub fn filter1(self) -> &'a mut W

Digital filter enabled and filtering capability up to 1 tI2CCLK

pub fn filter2(self) -> &'a mut W

Digital filter enabled and filtering capability up to 2 tI2CCLK

pub fn filter3(self) -> &'a mut W

Digital filter enabled and filtering capability up to 3 tI2CCLK

pub fn filter4(self) -> &'a mut W

Digital filter enabled and filtering capability up to 4 tI2CCLK

pub fn filter5(self) -> &'a mut W

Digital filter enabled and filtering capability up to 5 tI2CCLK

pub fn filter6(self) -> &'a mut W

Digital filter enabled and filtering capability up to 6 tI2CCLK

pub fn filter7(self) -> &'a mut W

Digital filter enabled and filtering capability up to 7 tI2CCLK

pub fn filter8(self) -> &'a mut W

Digital filter enabled and filtering capability up to 8 tI2CCLK

pub fn filter9(self) -> &'a mut W

Digital filter enabled and filtering capability up to 9 tI2CCLK

pub fn filter10(self) -> &'a mut W

Digital filter enabled and filtering capability up to 10 tI2CCLK

pub fn filter11(self) -> &'a mut W

Digital filter enabled and filtering capability up to 11 tI2CCLK

pub fn filter12(self) -> &'a mut W

Digital filter enabled and filtering capability up to 12 tI2CCLK

pub fn filter13(self) -> &'a mut W

Digital filter enabled and filtering capability up to 13 tI2CCLK

pub fn filter14(self) -> &'a mut W

Digital filter enabled and filtering capability up to 14 tI2CCLK

pub fn filter15(self) -> &'a mut W

Digital filter enabled and filtering capability up to 15 tI2CCLK

pub fn bits(self, value: u8) -> &'a mut W

Writes raw bits to the field