Module stm32f4::stm32f413::adc1::cr1

control register 1

Structs

AWDCHR
Value of the field

DISCNUMR
Value of the field

R
Value read from the register

W
Value to write to the register

_AWDCHW
Proxy

_AWDENW
Proxy

_AWDIEW
Proxy

_AWDSGLW
Proxy

_DISCENW
Proxy

_DISCNUMW
Proxy

_EOCIEW
Proxy

_JAUTOW
Proxy

_JAWDENW
Proxy

_JDISCENW
Proxy

_JEOCIEW
Proxy

_OVRIEW
Proxy

_RESW
Proxy

_SCANW
Proxy

Enums

AWDENR
Possible values of the field AWDEN

AWDENW
Values that can be written to the field AWDEN

AWDIER
Possible values of the field AWDIE

AWDIEW
Values that can be written to the field AWDIE

AWDSGLR
Possible values of the field AWDSGL

AWDSGLW
Values that can be written to the field AWDSGL

DISCENR
Possible values of the field DISCEN

DISCENW
Values that can be written to the field DISCEN

EOCIER
Possible values of the field EOCIE

EOCIEW
Values that can be written to the field EOCIE

JAUTOR
Possible values of the field JAUTO

JAUTOW
Values that can be written to the field JAUTO

JAWDENR
Possible values of the field JAWDEN

JAWDENW
Values that can be written to the field JAWDEN

JDISCENR
Possible values of the field JDISCEN

JDISCENW
Values that can be written to the field JDISCEN

JEOCIER
Possible values of the field JEOCIE

JEOCIEW
Values that can be written to the field JEOCIE

OVRIER
Possible values of the field OVRIE

OVRIEW
Values that can be written to the field OVRIE

RESR
Possible values of the field RES

RESW
Values that can be written to the field RES

SCANR
Possible values of the field SCAN

SCANW
Values that can be written to the field SCAN