Enum stm32f4::stm32f413::dma1::s0fcr::FSR

pub enum FSR {
    FIRSTQUARTER,
    SECONDQUARTER,
    THIRDQUARTER,
    FOURTHQUARTER,
    EMPTY,
    FULL,
    _Reserved(u8),
}

Possible values of the field FS

Variants
FIRSTQUARTER

0 < fifo_level < 1/4

SECONDQUARTER

1/4 <= fifo_level < 1/2

THIRDQUARTER

1/2 <= fifo_level < 3/4

FOURTHQUARTER

3/4 <= fifo_level < full

EMPTY

FIFO is empty

FULL

FIFO is full

_Reserved(u8)

Reserved

Methods

impl FSR

pub fn bits(&self) -> u8

Value of the field as raw bits

pub fn is_first_quarter(&self) -> bool

Checks if the value of the field is FIRSTQUARTER

pub fn is_second_quarter(&self) -> bool

Checks if the value of the field is SECONDQUARTER

pub fn is_third_quarter(&self) -> bool

Checks if the value of the field is THIRDQUARTER

pub fn is_fourth_quarter(&self) -> bool

Checks if the value of the field is FOURTHQUARTER

pub fn is_empty(&self) -> bool

Checks if the value of the field is EMPTY

pub fn is_full(&self) -> bool

Checks if the value of the field is FULL