Module stm32f4::stm32f413::dma1::s7cr

stream x configuration register

Structs

ACKR
Value of the field

CHSELR
Value of the field

R
Value read from the register

W
Value to write to the register

_ACKW
Proxy

_CHSELW
Proxy

_CIRCW
Proxy

_CTW
Proxy

_DBMW
Proxy

_DIRW
Proxy

_DMEIEW
Proxy

_ENW
Proxy

_HTIEW
Proxy

_MBURSTW
Proxy

_MINCW
Proxy

_MSIZEW
Proxy

_PBURSTW
Proxy

_PFCTRLW
Proxy

_PINCOSW
Proxy

_PINCW
Proxy

_PLW
Proxy

_PSIZEW
Proxy

_TCIEW
Proxy

_TEIEW
Proxy

Enums

CIRCR
Possible values of the field CIRC

CIRCW
Values that can be written to the field CIRC

CTR
Possible values of the field CT

CTW
Values that can be written to the field CT

DBMR
Possible values of the field DBM

DBMW
Values that can be written to the field DBM

DIRR
Possible values of the field DIR

DIRW
Values that can be written to the field DIR

DMEIER
Possible values of the field DMEIE

DMEIEW
Values that can be written to the field DMEIE

ENR
Possible values of the field EN

ENW
Values that can be written to the field EN

HTIER
Possible values of the field HTIE

HTIEW
Values that can be written to the field HTIE

MBURSTR
Possible values of the field MBURST

MBURSTW
Values that can be written to the field MBURST

MINCR
Possible values of the field MINC

MINCW
Values that can be written to the field MINC

MSIZER
Possible values of the field MSIZE

MSIZEW
Values that can be written to the field MSIZE

PFCTRLR
Possible values of the field PFCTRL

PFCTRLW
Values that can be written to the field PFCTRL

PINCOSR
Possible values of the field PINCOS

PINCOSW
Values that can be written to the field PINCOS

PLR
Possible values of the field PL

PLW
Values that can be written to the field PL

TCIER
Possible values of the field TCIE

TCIEW
Values that can be written to the field TCIE

TEIER
Possible values of the field TEIE

TEIEW
Values that can be written to the field TEIE

Type Definitions

PBURSTR
Possible values of the field PBURST

PBURSTW
Values that can be written to the field PBURST

PINCR
Possible values of the field PINC

PINCW
Values that can be written to the field PINC

PSIZER
Possible values of the field PSIZE

PSIZEW
Values that can be written to the field PSIZE