Enum stm32f4::stm32f413::gpiob::afrl::AFRL7R

pub enum AFRL7R {
    AF0,
    AF1,
    AF2,
    AF3,
    AF4,
    AF5,
    AF6,
    AF7,
    AF8,
    AF9,
    AF10,
    AF11,
    AF12,
    AF13,
    AF14,
    AF15,
}

Possible values of the field AFRL7

Variants
AF0

AF0

AF1

AF1

AF2

AF2

AF3

AF3

AF4

AF4

AF5

AF5

AF6

AF6

AF7

AF7

AF8

AF8

AF9

AF9

AF10

AF10

AF11

AF11

AF12

AF12

AF13

AF13

AF14

AF14

AF15

AF15

Methods

impl AFRL7R

pub fn bits(&self) -> u8

Value of the field as raw bits

pub fn is_af0(&self) -> bool

Checks if the value of the field is AF0

pub fn is_af1(&self) -> bool

Checks if the value of the field is AF1

pub fn is_af2(&self) -> bool

Checks if the value of the field is AF2

pub fn is_af3(&self) -> bool

Checks if the value of the field is AF3

pub fn is_af4(&self) -> bool

Checks if the value of the field is AF4

pub fn is_af5(&self) -> bool

Checks if the value of the field is AF5

pub fn is_af6(&self) -> bool

Checks if the value of the field is AF6

pub fn is_af7(&self) -> bool

Checks if the value of the field is AF7

pub fn is_af8(&self) -> bool

Checks if the value of the field is AF8

pub fn is_af9(&self) -> bool

Checks if the value of the field is AF9

pub fn is_af10(&self) -> bool

Checks if the value of the field is AF10

pub fn is_af11(&self) -> bool

Checks if the value of the field is AF11

pub fn is_af12(&self) -> bool

Checks if the value of the field is AF12

pub fn is_af13(&self) -> bool

Checks if the value of the field is AF13

pub fn is_af14(&self) -> bool

Checks if the value of the field is AF14

pub fn is_af15(&self) -> bool

Checks if the value of the field is AF15