Module stm32f4::stm32f413::gpiob::odr

GPIO port output data register

Structs

R
Value read from the register

W
Value to write to the register

_ODR0W
Proxy

_ODR1W
Proxy

_ODR2W
Proxy

_ODR3W
Proxy

_ODR4W
Proxy

_ODR5W
Proxy

_ODR6W
Proxy

_ODR7W
Proxy

_ODR8W
Proxy

_ODR9W
Proxy

_ODR10W
Proxy

_ODR11W
Proxy

_ODR12W
Proxy

_ODR13W
Proxy

_ODR14W
Proxy

_ODR15W
Proxy

Enums

ODR15R
Possible values of the field ODR15

ODR15W
Values that can be written to the field ODR15

Type Definitions

ODR0R
Possible values of the field ODR0

ODR0W
Values that can be written to the field ODR0

ODR1R
Possible values of the field ODR1

ODR1W
Values that can be written to the field ODR1

ODR2R
Possible values of the field ODR2

ODR2W
Values that can be written to the field ODR2

ODR3R
Possible values of the field ODR3

ODR3W
Values that can be written to the field ODR3

ODR4R
Possible values of the field ODR4

ODR4W
Values that can be written to the field ODR4

ODR5R
Possible values of the field ODR5

ODR5W
Values that can be written to the field ODR5

ODR6R
Possible values of the field ODR6

ODR6W
Values that can be written to the field ODR6

ODR7R
Possible values of the field ODR7

ODR7W
Values that can be written to the field ODR7

ODR8R
Possible values of the field ODR8

ODR8W
Values that can be written to the field ODR8

ODR9R
Possible values of the field ODR9

ODR9W
Values that can be written to the field ODR9

ODR10R
Possible values of the field ODR10

ODR10W
Values that can be written to the field ODR10

ODR11R
Possible values of the field ODR11

ODR11W
Values that can be written to the field ODR11

ODR12R
Possible values of the field ODR12

ODR12W
Values that can be written to the field ODR12

ODR13R
Possible values of the field ODR13

ODR13W
Values that can be written to the field ODR13

ODR14R
Possible values of the field ODR14

ODR14W
Values that can be written to the field ODR14