Module stm32f4::stm32f413::i2c1::sr1

Status register 1

Structs

ADD10R
Value of the field

R
Value read from the register

W
Value to write to the register

_AFW
Proxy

_ARLOW
Proxy

_BERRW
Proxy

_OVRW
Proxy

_PECERRW
Proxy

_SMBALERTW
Proxy

_TIMEOUTW
Proxy

Enums

ADDRR
Possible values of the field ADDR

AFR
Possible values of the field AF

AFW
Values that can be written to the field AF

ARLOR
Possible values of the field ARLO

ARLOW
Values that can be written to the field ARLO

BERRR
Possible values of the field BERR

BERRW
Values that can be written to the field BERR

BTFR
Possible values of the field BTF

OVRR
Possible values of the field OVR

OVRW
Values that can be written to the field OVR

PECERRR
Possible values of the field PECERR

PECERRW
Values that can be written to the field PECERR

RXNER
Possible values of the field RxNE

SBR
Possible values of the field SB

SMBALERTR
Possible values of the field SMBALERT

SMBALERTW
Values that can be written to the field SMBALERT

STOPFR
Possible values of the field STOPF

TIMEOUTR
Possible values of the field TIMEOUT

TIMEOUTW
Values that can be written to the field TIMEOUT

TXER
Possible values of the field TxE