Module stm32f4::stm32f413::rcc::ahb2lpenr

AHB2 peripheral clock enable in low power mode register

Structs

R
Value read from the register

W
Value to write to the register

_FSMCLPENW
Proxy

_OTGFSLPENW
Proxy

_QSPILPENW
Proxy

_RNGLPENW
Proxy

Enums

OTGFSLPENR
Possible values of the field OTGFSLPEN

OTGFSLPENW
Values that can be written to the field OTGFSLPEN

Type Definitions

FSMCLPENR
Possible values of the field FSMCLPEN

FSMCLPENW
Values that can be written to the field FSMCLPEN

QSPILPENR
Possible values of the field QSPILPEN

QSPILPENW
Values that can be written to the field QSPILPEN

RNGLPENR
Possible values of the field RNGLPEN

RNGLPENW
Values that can be written to the field RNGLPEN