Module stm32f4::stm32f413::rcc::dckcfgr

Dedicated Clocks Configuration Register

Structs

R
Value read from the register

W
Value to write to the register

_CKDFSDM1ASELW
Proxy

_CKDFSDM1SELW
Proxy

_CKDFSDM2ASELW
Proxy

_I2S1SRCW
Proxy

_I2S2SRCW
Proxy

_PLLDIVRW
Proxy

_PLLI2SDIVRW
Proxy

_SAI1ASRCW
Proxy

_SAI1BSRCW
Proxy

_TIMPREW
Proxy

Enums

CKDFSDM1ASELR
Possible values of the field CKDFSDM1ASEL

CKDFSDM1SELR
Possible values of the field CKDFSDM1SEL

CKDFSDM1ASELW
Values that can be written to the field CKDFSDM1ASEL

CKDFSDM1SELW
Values that can be written to the field CKDFSDM1SEL

I2S1SRCR
Possible values of the field I2S1SRC

I2S1SRCW
Values that can be written to the field I2S1SRC

PLLDIVRR
Possible values of the field PLLDIVR

PLLDIVRW
Values that can be written to the field PLLDIVR

PLLI2SDIVRR
Possible values of the field PLLI2SDIVR

PLLI2SDIVRW
Values that can be written to the field PLLI2SDIVR

SAI1ASRCR
Possible values of the field SAI1ASRC

SAI1BSRCR
Possible values of the field SAI1BSRC

SAI1ASRCW
Values that can be written to the field SAI1ASRC

SAI1BSRCW
Values that can be written to the field SAI1BSRC

TIMPRER
Possible values of the field TIMPRE

TIMPREW
Values that can be written to the field TIMPRE

Type Definitions

CKDFSDM2ASELR
Possible values of the field CKDFSDM2ASEL

CKDFSDM2ASELW
Values that can be written to the field CKDFSDM2ASEL

I2S2SRCR
Possible values of the field I2S2SRC

I2S2SRCW
Values that can be written to the field I2S2SRC