Module stm32f4::stm32f413::spi5::cr1

control register 1

Structs

R
Value read from the register

SSIR
Value of the field

W
Value to write to the register

_BIDIMODEW
Proxy

_BIDIOEW
Proxy

_BRW
Proxy

_CPHAW
Proxy

_CPOLW
Proxy

_CRCENW
Proxy

_CRCNEXTW
Proxy

_DFFW
Proxy

_LSBFIRSTW
Proxy

_MSTRW
Proxy

_RXONLYW
Proxy

_SPEW
Proxy

_SSIW
Proxy

_SSMW
Proxy

Enums

BIDIMODER
Possible values of the field BIDIMODE

BIDIMODEW
Values that can be written to the field BIDIMODE

BIDIOER
Possible values of the field BIDIOE

BIDIOEW
Values that can be written to the field BIDIOE

BRR
Possible values of the field BR

BRW
Values that can be written to the field BR

CPHAR
Possible values of the field CPHA

CPHAW
Values that can be written to the field CPHA

CPOLR
Possible values of the field CPOL

CPOLW
Values that can be written to the field CPOL

CRCENR
Possible values of the field CRCEN

CRCENW
Values that can be written to the field CRCEN

CRCNEXTR
Possible values of the field CRCNEXT

CRCNEXTW
Values that can be written to the field CRCNEXT

DFFR
Possible values of the field DFF

DFFW
Values that can be written to the field DFF

LSBFIRSTR
Possible values of the field LSBFIRST

LSBFIRSTW
Values that can be written to the field LSBFIRST

MSTRR
Possible values of the field MSTR

MSTRW
Values that can be written to the field MSTR

RXONLYR
Possible values of the field RXONLY

RXONLYW
Values that can be written to the field RXONLY

SPER
Possible values of the field SPE

SPEW
Values that can be written to the field SPE

SSMR
Possible values of the field SSM

SSMW
Values that can be written to the field SSM