Module stm32f4::stm32f413::spi5::i2scfgr

I2S configuration register

Structs

R
Value read from the register

W
Value to write to the register

_CHLENW
Proxy

_CKPOLW
Proxy

_DATLENW
Proxy

_I2SMODW
Proxy

_I2SEW
Proxy

_I2SCFGW
Proxy

_I2SSTDW
Proxy

_PCMSYNCW
Proxy

Enums

CHLENR
Possible values of the field CHLEN

CHLENW
Values that can be written to the field CHLEN

CKPOLR
Possible values of the field CKPOL

CKPOLW
Values that can be written to the field CKPOL

DATLENR
Possible values of the field DATLEN

DATLENW
Values that can be written to the field DATLEN

I2SMODR
Possible values of the field I2SMOD

I2SER
Possible values of the field I2SE

I2SCFGR
Possible values of the field I2SCFG

I2SSTDR
Possible values of the field I2SSTD

I2SMODW
Values that can be written to the field I2SMOD

I2SEW
Values that can be written to the field I2SE

I2SCFGW
Values that can be written to the field I2SCFG

I2SSTDW
Values that can be written to the field I2SSTD

PCMSYNCR
Possible values of the field PCMSYNC

PCMSYNCW
Values that can be written to the field PCMSYNC