Trait stm32f4::Resettable

pub trait Resettable: RegisterSpec {
    fn reset_value() -> Self::Ux;
}

Reset value of the register.

This value is the initial value for the write method. It can also be directly written to the register by using the reset method.

Required methods

fn reset_value() -> Self::Ux

Reset value of the register.