Crate stm32f4

Peripheral access API for STM32F4 microcontrollers (generated using svd2rust 0.19.0)

You can find an overview of the API here: svd2rust/#peripheral-api

For more details see the README here: stm32-rs

This crate supports all STM32F4 devices; for the complete list please see: stm32f4

Due to doc build limitations, not all devices may be shown on docs.rs; a representative few have been selected instead. For a complete list of available registers and fields see: stm32-rs Device Coverage

Modules

stm32f411

Peripheral access API for STM32F411 microcontrollers (generated using svd2rust v0.19.0 ( ))

Structs

FieldReader

Field reader.

R

Register reader.

Reg

This structure provides volatile access to registers.

W

Register writer.

Traits

Readable

Trait implemented by readable registers to enable the read method.

RegisterSpec

Raw register type

Resettable

Reset value of the register.

Writable

Trait implemented by writeable registers.